@@ -1,21 +1,30 @@
 {
   "name": "dotenv",
-  "version": "8.6.0",
+  "version": "16.4.5",
   "description": "Loads environment variables from .env file",
   "main": "lib/main.js",
+  "types": "lib/main.d.ts",
   "exports": {
-    ".": "./lib/main.js",
+    ".": {
+      "types": "./lib/main.d.ts",
+      "require": "./lib/main.js",
+      "default": "./lib/main.js"
+    },
     "./config": "./config.js",
+    "./config.js": "./config.js",
+    "./lib/env-options": "./lib/env-options.js",
+    "./lib/env-options.js": "./lib/env-options.js",
+    "./lib/cli-options": "./lib/cli-options.js",
+    "./lib/cli-options.js": "./lib/cli-options.js",
     "./package.json": "./package.json"
   },
-  "types": "types/index.d.ts",
   "scripts": {
-    "flow": "flow",
-    "dtslint": "dtslint types",
+    "dts-check": "tsc --project tests/types/tsconfig.json",
     "lint": "standard",
-    "postlint": "standard-markdown",
-    "pretest": "npm run lint && npm run dtslint",
-    "test": "tap tests/*.js --100",
+    "lint-readme": "standard-markdown",
+    "pretest": "npm run lint && npm run dts-check",
+    "test": "tap tests/*.js --100 -Rspec",
+    "test:coverage": "tap --coverage-report=lcov",
     "prerelease": "npm test",
     "release": "standard-version"
   },
@@ -23,6 +32,7 @@
     "type": "git",
     "url": "git://github.com/motdotla/dotenv.git"
   },
+  "funding": "https://dotenvx.com",
   "keywords": [
     "dotenv",
     "env",
@@ -35,22 +45,21 @@
   "readmeFilename": "README.md",
   "license": "BSD-2-Clause",
   "devDependencies": {
-    "decache": "^4.5.1",
-    "dtslint": "^0.9.8",
-    "flow-bin": "^0.109.0",
-    "sinon": "^7.5.0",
-    "standard": "^13.1.0",
-    "standard-markdown": "^5.1.0",
-    "standard-version": "^7.0.0",
-    "tap": "^14.7.0"
+    "@definitelytyped/dtslint": "^0.0.133",
+    "@types/node": "^18.11.3",
+    "decache": "^4.6.1",
+    "sinon": "^14.0.1",
+    "standard": "^17.0.0",
+    "standard-markdown": "^7.1.0",
+    "standard-version": "^9.5.0",
+    "tap": "^16.3.0",
+    "tar": "^6.1.11",
+    "typescript": "^4.8.4"
   },
-  "dependencies": {},
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
-  "standard": {
-    "ignore": [
-      "flow-typed/"
-    ]
+  "browser": {
+    "fs": false
   }
 }
