@@ -1,5 +1,6 @@
 "use strict";
 
+require("@babel/register");
 const express = require("express");
 const cors = require("cors");
 
@@ -29,7 +30,7 @@ app.get("/", (req, res) => {
   res.send("Welcome to the KFG Backend API");
 });
 
-app.get('/favicon.ico', (req, res) => res.status(204));
+app.get("/favicon.ico", (req, res) => res.status(204));
 
 /** Handle 404 errors -- this matches everything */
 app.use(function (req, res, next) {
