@@ -9,20 +9,17 @@ const { sqlForPartialUpdate } = require("../helpers/sql");
 class skill {
   /** Create a skill (from data), update db, return new skill data.
    *
-   * data should be { title, salary, equity, employeeHandle }
+   * data should be { skill_name, description }
    *
-   * Returns { id, title, salary, equity, employeeHandle }
+   * Returns { id, skill_name, description }
    **/
 
   static async create(data) {
     const result = await db.query(
-      `INSERT INTO skills (title,
-                             salary,
-                             equity,
-                             employee_handle)
-           VALUES ($1, $2, $3, $4)
-           RETURNING id, title, salary, equity, employee_handle AS "employeeHandle"`,
-      [data.title, data.salary, data.equity, data.employeeHandle]
+      `INSERT INTO skills (skill_name, description)
+           VALUES ($1, $2)
+           RETURNING id, skill_name, description"`,
+      [data.skill_name, data.description]
     );
     let skill = result.rows[0];
 
@@ -32,68 +29,33 @@ class skill {
   /** Find all skills (optional filter on searchFilters).
    *
    * searchFilters (all optional):
-   * - minSalary
-   * - hasEquity (true returns only skills with equity > 0, other values ignored)
-   * - title (will find case-insensitive, partial matches)
+   * - skill_name (will find case-insensitive, partial matches)
    *
-   * Returns [{ id, title, salary, equity, employeeHandle, employeeName }, ...]
+   * Returns [{ id, skill_name, description }, ...]
    * */
 
-  static async findAll({ minSalary, hasEquity, title } = {}) {
-    let query = `SELECT j.id,
-                        j.title,
-                        j.salary,
-                        j.equity,
-                        j.employee_handle AS "employeeHandle",
-                        c.name AS "employeeName"
-                 FROM skills j 
-                   LEFT JOIN employees AS c ON c.handle = j.employee_handle`;
+  static async findAll({ skill_name, description } = {}) {
+    let query = `SELECT s.skill_name,
+                        s.description,  
+                 FROM skills s `;
     let whereExpressions = [];
     let queryValues = [];
 
-    // For each possible search term, add to whereExpressions and
-    // queryValues so we can generate the right SQL
-
-    if (minSalary !== undefined) {
-      queryValues.push(minSalary);
-      whereExpressions.push(`salary >= $${queryValues.length}`);
-    }
-
-    if (hasEquity === true) {
-      whereExpressions.push(`equity > 0`);
-    }
-
-    if (title !== undefined) {
-      queryValues.push(`%${title}%`);
-      whereExpressions.push(`title ILIKE $${queryValues.length}`);
-    }
-
-    if (whereExpressions.length > 0) {
-      query += " WHERE " + whereExpressions.join(" AND ");
-    }
-
     // Finalize query and return results
 
-    query += " ORDER BY title";
+    query += " ORDER BY skill_name";
     const skillsRes = await db.query(query, queryValues);
     return skillsRes.rows;
   }
 
-  /** Given a skill id, return data about skill.
-   *
-   * Returns { id, title, salary, equity, employeeHandle, employee }
-   *   where employee is { handle, name, description, numEmployees, logoUrl }
-   *
-   * Throws NotFoundError if not found.
-   **/
+  // /** Given a skill id, return data about skill.
+
 
   static async get(id) {
     const skillRes = await db.query(
       `SELECT id,
-                  title,
-                  salary,
-                  equity,
-                  employee_handle AS "employeeHandle"
+                  skill_name AS "skillName",
+                  description,
            FROM skills
            WHERE id = $1`,
       [id]
@@ -104,11 +66,8 @@ class skill {
     if (!skill) throw new NotFoundError(`No skill: ${id}`);
 
     const employeesRes = await db.query(
-      `SELECT handle,
-                  name,
-                  description,
-                  num_employees AS "numEmployees",
-                  logo_url AS "logoUrl"
+      `SELECT 
+      
            FROM employees
            WHERE handle = $1`,
       [skill.employeeHandle]
